Парсинг сгруппирован в сервисе
